+++ typescript/polyglotpkg/src/strategies/powershell.ts
@@ -93,17 +93,22 @@ export class PowershellStrategy extends BaseStrategy {
 
         deps.split(/\r?\n/).forEach(line => {
             if (line.indexOf("Import-Module") !== -1 && line.indexOf("#@ignore") === -1 && !line.startsWith("#")) {
-                const module: string = line.replace("Import-Module", "").trim();
+                const module: string = line.replace("Import-Module", "").replace("-Name", "").replace(";", "").trim();
                 modules.push(module);
             }
         });
 
         this.logger.info(`Powershell modules included: ${modules}`);
 
+        const moduleNames = modules.join(",");
+        const securityProtocolArg = polyglotJson.platform.protocolType ? `[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::${polyglotJson.platform.protocolType};` : "";
+        const command = `${securityProtocolArg} Save-Module -Name ${moduleNames} -Path '${modulesPath}' -Repository PSGallery`;
+        
+        const args = polyglotJson.platform.protocolType ? [`-c "${command}"`] : [command];
         if (modules.length > 0) {
             fs.ensureDirSync(modulesPath);
             this.logger.info(`Downloading and saving dependencies in "${modulesPath}..."`);
-            await run("pwsh", ["-c", "Save-Module", "-Name", `"${modules.toString()}"`, "-Path", `"${modulesPath}"`, "-Repository PSGallery"]);
+            await run("pwsh", args);
         } else {
             this.logger.info("No change in dependencies. Skipping installation...");
         }